Fix TUI not displaying data after options change
https://github.com/kubecost/kubectl-cost/pull/41 introduced
an error with the TUI because the TUI was not providing
the standard set of options and was instead relying on
this old logic to set a namespace. The TUI now has the correct
flags and defaults for the KubeOptions, mitigating the issue.

Tested locally against a cluster running the nightly deployment.