Add support for standard kubectl tracing
I am having a problem with the portforwarding into the pod.
Standard approach for debugging these kinds of issues is to pass `-v=9` argument to `kubectl` so I see the actual API traffic dumps in the logs.
But `kubectl cost` doesn't support that argument, so there is no way to turn the K8S API client library tracing on.
Hence, I have added support for those arguments.